Separate OfaRuntimeWrapper impls into separate files
