--- crates/relayer-framework/src/base/one_for_all/impls/runtime.rs
@@ -1,96 +1,0 @@
-use async_trait::async_trait;
-use core::future::Future;
-use core::time::Duration;
-
-use crate::base::core::traits::error::HasErrorType;
-use crate::base::one_for_all::traits::runtime::{LogLevel, OfaRuntime, OfaRuntimeWrapper};
-use crate::base::runtime::traits::log::{
-    HasLogger, LevelDebug, LevelError, LevelInfo, LevelTrace, LevelWarn,
-};
-use crate::base::runtime::traits::sleep::CanSleep;
-use crate::base::runtime::traits::spawn::{HasSpawner, Spawner};
-use crate::base::runtime::traits::time::{HasTime, Time};
-use crate::std_prelude::*;
-
-pub struct OfaTime<Runtime: OfaRuntime> {
-    pub time: Runtime::Time,
-}
-
-impl<Runtime: OfaRuntime> HasErrorType for OfaRuntimeWrapper<Runtime> {
-    type Error = Runtime::Error;
-}
-
-#[async_trait]
-impl<Runtime: OfaRuntime> CanSleep for OfaRuntimeWrapper<Runtime> {
-    async fn sleep(&self, duration: Duration) {
-        self.runtime.sleep(duration).await
-    }
-}
-
-impl<Runtime: OfaRuntime> HasSpawner for OfaRuntimeWrapper<Runtime> {
-    type Spawner = Self;
-
-    fn spawner(&self) -> Self::Spawner {
-        self.clone()
-    }
-}
-
-impl<Runtime: OfaRuntime> Spawner for OfaRuntimeWrapper<Runtime> {
-    fn spawn<F>(&self, task: F)
-    where
-        F: Future + Send + 'static,
-        F::Output: Send + 'static,
-    {
-        self.runtime.spawn(task)
-    }
-}
-
-impl<Runtime: OfaRuntime> HasTime for OfaRuntimeWrapper<Runtime> {
-    type Time = OfaTime<Runtime>;
-
-    fn now(&self) -> Self::Time {
-        let time = self.runtime.now();
-        OfaTime { time }
-    }
-}
-
-impl<Runtime: OfaRuntime> Time for OfaTime<Runtime> {
-    fn duration_since(&self, other: &Self) -> Duration {
-        Runtime::duration_since(&self.time, &other.time)
-    }
-}
-
-#[async_trait]
-impl<Runtime: OfaRuntime> HasLogger<LevelError> for OfaRuntimeWrapper<Runtime> {
-    async fn log(&self, _level: LevelError, message: &str) {
-        self.runtime.log(LogLevel::Error, message).await;
-    }
-}
-
-#[async_trait]
-impl<Runtime: OfaRuntime> HasLogger<LevelWarn> for OfaRuntimeWrapper<Runtime> {
-    async fn log(&self, _level: LevelWarn, message: &str) {
-        self.runtime.log(LogLevel::Warn, message).await;
-    }
-}
-
-#[async_trait]
-impl<Runtime: OfaRuntime> HasLogger<LevelInfo> for OfaRuntimeWrapper<Runtime> {
-    async fn log(&self, _level: LevelInfo, message: &str) {
-        self.runtime.log(LogLevel::Info, message).await;
-    }
-}
-
-#[async_trait]
-impl<Runtime: OfaRuntime> HasLogger<LevelDebug> for OfaRuntimeWrapper<Runtime> {
-    async fn log(&self, _level: LevelDebug, message: &str) {
-        self.runtime.log(LogLevel::Debug, message).await;
-    }
-}
-
-#[async_trait]
-impl<Runtime: OfaRuntime> HasLogger<LevelTrace> for OfaRuntimeWrapper<Runtime> {
-    async fn log(&self, _level: LevelTrace, message: &str) {
-        self.runtime.log(LogLevel::Trace, message).await;
-    }
-}
