+++ crates/relayer-framework/src/base/one_for_all/impls/runtime/sleep.rs
@@ -1,0 +1,13 @@
+use async_trait::async_trait;
+use core::time::Duration;
+
+use crate::base::one_for_all::traits::runtime::{OfaRuntime, OfaRuntimeWrapper};
+use crate::base::runtime::traits::sleep::CanSleep;
+use crate::std_prelude::*;
+
+#[async_trait]
+impl<Runtime: OfaRuntime> CanSleep for OfaRuntimeWrapper<Runtime> {
+    async fn sleep(&self, duration: Duration) {
+        self.runtime.sleep(duration).await
+    }
+}
