+++ crates/relayer-framework/src/base/runtime/traits/time.rs
@@ -2,12 +2,10 @@ use core::time::Duration;
 
 use crate::base::core::traits::sync::Async;
 
-pub trait Time: Async {
-    fn duration_since(&self, other: &Self) -> Duration;
-}
-
 pub trait HasTime: Async {
-    type Time: Time;
+    type Time: Async;
 
     fn now(&self) -> Self::Time;
+
+    fn duration_since(current_time: &Self::Time, other_time: &Self::Time) -> Duration;
 }
